Reduced: project builder features
